backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
passlib[bcrypt]
bcrypt==3.2.0
pyjwt
python-dotenv
python-multipart
requests